zstd: use int positions throughout
Buffer positions were longs left over from the Unsafe days, when a position
was an absolute machine address. Every position is now an index into a
byte[], so it is an int by definition; the long-to-int casts at the access
sites (over 300 of them) are gone with it, along with the possibility of a
silent truncation.

Values that are genuinely 64-bit stay long: bit containers, hashes, the
FSE normalization arithmetic, the frame content size and dictionary id,
and the public getDecompressedSize API. The native entry points are
untouched.

One deliberate exception: the main loop of DoubleFastBlockCompressor keeps
its loop-carried positions in long locals behind the int signature. With
int loop variables C2 generates measurably slower code for that method
(compression about 10% slower on silesia/x-ray, 5% overall), and the cause
is not any of the usual loop optimizations. The values are bounded by the
int inputs, so the remaining casts there cannot truncate; a comment says so.

Compressed output is byte-identical to the Unsafe implementation.
Decompression throughput relative to Unsafe: geomean 0.95 (was 0.92);
compression unchanged at parity.

+++ src/main/java/io/airlift/compress/v3/zstd/BitInputStream.java
@@ -30,28 +30,28 @@ final class BitInputStream
     {
     }
 
-    public static boolean isEndOfStream(long startAddress, long currentAddress, int bitsConsumed)
+    public static boolean isEndOfStream(int startAddress, int currentAddress, int bitsConsumed)
     {
         return startAddress == currentAddress && bitsConsumed == Long.SIZE;
     }
 
-    private static long readTail(byte[] inputBase, long inputAddress, int inputSize)
+    private static long readTail(byte[] inputBase, int inputAddress, int inputSize)
     {
-        long bits = inputBase[(int) inputAddress] & 0xFF;
+        long bits = inputBase[inputAddress] & 0xFF;
 
         switch (inputSize) {
             case 7:
-                bits |= (inputBase[(int) (inputAddress + 6)] & 0xFFL) << 48;
+                bits |= (inputBase[inputAddress + 6] & 0xFFL) << 48;
             case 6:
-                bits |= (inputBase[(int) (inputAddress + 5)] & 0xFFL) << 40;
+                bits |= (inputBase[inputAddress + 5] & 0xFFL) << 40;
             case 5:
-                bits |= (inputBase[(int) (inputAddress + 4)] & 0xFFL) << 32;
+                bits |= (inputBase[inputAddress + 4] & 0xFFL) << 32;
             case 4:
-                bits |= (inputBase[(int) (inputAddress + 3)] & 0xFFL) << 24;
+                bits |= (inputBase[inputAddress + 3] & 0xFFL) << 24;
             case 3:
-                bits |= (inputBase[(int) (inputAddress + 2)] & 0xFFL) << 16;
+                bits |= (inputBase[inputAddress + 2] & 0xFFL) << 16;
             case 2:
-                bits |= (inputBase[(int) (inputAddress + 1)] & 0xFFL) << 8;
+                bits |= (inputBase[inputAddress + 1] & 0xFFL) << 8;
         }
 
         return bits;
@@ -78,13 +78,13 @@ final class BitInputStream
     static class Initializer
     {
         private final byte[] inputBase;
-        private final long startAddress;
-        private final long endAddress;
+        private final int startAddress;
+        private final int endAddress;
         private long bits;
-        private long currentAddress;
+        private int currentAddress;
         private int bitsConsumed;
 
-        public Initializer(byte[] inputBase, long startAddress, long endAddress)
+        public Initializer(byte[] inputBase, int startAddress, int endAddress)
         {
             this.inputBase = inputBase;
             this.startAddress = startAddress;
@@ -96,7 +96,7 @@ final class BitInputStream
             return bits;
         }
 
-        public long getCurrentAddress()
+        public int getCurrentAddress()
         {
             return currentAddress;
         }
@@ -110,15 +110,15 @@ final class BitInputStream
         {
             verify(endAddress - startAddress >= 1, startAddress, "Bitstream is empty");
 
-            int lastByte = inputBase[(int) (endAddress - 1)] & 0xFF;
+            int lastByte = inputBase[endAddress - 1] & 0xFF;
             verify(lastByte != 0, endAddress, "Bitstream end mark not present");
 
             bitsConsumed = SIZE_OF_LONG - highestBit(lastByte);
 
-            int inputSize = (int) (endAddress - startAddress);
+            int inputSize = endAddress - startAddress;
             if (inputSize >= SIZE_OF_LONG) {  /* normal case */
                 currentAddress = endAddress - SIZE_OF_LONG;
-                bits = (long) Mem.LONG_LE.get(inputBase, (int) currentAddress);
+                bits = (long) Mem.LONG_LE.get(inputBase, currentAddress);
             }
             else {
                 currentAddress = startAddress;
@@ -132,13 +132,13 @@ final class BitInputStream
     static final class Loader
     {
         private final byte[] inputBase;
-        private final long startAddress;
+        private final int startAddress;
         private long bits;
-        private long currentAddress;
+        private int currentAddress;
         private int bitsConsumed;
         private boolean overflow;
 
-        public Loader(byte[] inputBase, long startAddress, long currentAddress, long bits, int bitsConsumed)
+        public Loader(byte[] inputBase, int startAddress, int currentAddress, long bits, int bitsConsumed)
         {
             this.inputBase = inputBase;
             this.startAddress = startAddress;
@@ -152,7 +152,7 @@ final class BitInputStream
             return bits;
         }
 
-        public long getCurrentAddress()
+        public int getCurrentAddress()
         {
             return currentAddress;
         }
@@ -182,21 +182,21 @@ final class BitInputStream
             if (currentAddress >= startAddress + SIZE_OF_LONG) {
                 if (bytes > 0) {
                     currentAddress -= bytes;
-                    bits = (long) Mem.LONG_LE.get(inputBase, (int) currentAddress);
+                    bits = (long) Mem.LONG_LE.get(inputBase, currentAddress);
                 }
                 bitsConsumed &= 0b111;
             }
             else if (currentAddress - bytes < startAddress) {
-                bytes = (int) (currentAddress - startAddress);
+                bytes = currentAddress - startAddress;
                 currentAddress = startAddress;
                 bitsConsumed -= bytes * SIZE_OF_LONG;
-                bits = (long) Mem.LONG_LE.get(inputBase, (int) startAddress);
+                bits = (long) Mem.LONG_LE.get(inputBase, startAddress);
                 return true;
             }
             else {
                 currentAddress -= bytes;
                 bitsConsumed -= bytes * SIZE_OF_LONG;
-                bits = (long) Mem.LONG_LE.get(inputBase, (int) currentAddress);
+                bits = (long) Mem.LONG_LE.get(inputBase, currentAddress);
             }
 
             return false;
